Move issue/severity table to submit_ce/ui/preflight (out of controller). [SUBMISSION-211] David
